+++ apps/mobile-wallet/src/navigation/InWalletNavigation.tsx
@@ -2,7 +2,7 @@ import Ionicons from '@expo/vector-icons/Ionicons'
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'
 import { StatusBar } from 'expo-status-bar'
 import { useTranslation } from 'react-i18next'
-import { useTheme } from 'styled-components'
+import { useTheme } from 'styled-components/native'
 
 import FooterMenu from '~/components/footers/FooterMenu'
 import EcosystemScreen from '~/features/ecosystem/EcosystemScreen'